instageo/model/requirements.txt
pytorch_lightning
torch
timm==1.0.15
einops
tensorboard
omegaconf
hydra-core
torchvision
huggingface_hub
ray[serve]
neptune
